Refactor Star into components
